% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_proportions.R
\name{actual_proportions}
\alias{actual_proportions}
\title{The proportion within each class AFTER class assignment (using max posterior rule)}
\usage{
actual_proportions(p)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}
}
\value{
The proportion within each class AFTER class assignment (using max posterior rule)
}
\description{
The proportion within each class AFTER class assignment (using max posterior rule)
}
\examples{
\dontrun{actual_proportions(p)}
}
