% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\docType{data}
\name{bmi}
\alias{bmi}
\title{Body Mass Index (BMI) repeated measures of 10,000 individuals in wide format.}
\format{A wide format data frame of simulated BMI values of 10,000 individuals.
\describe{
  \item{id}{Individual ID}
  \item{bmi}{Body mass index of the individual at times T1,T2, T3 and T4, in kg/m^2}
  \item{T}{Time of BMI measure, in years}
  \item{true_class}{Tag to identify the class the individual BMI data was simulated from}
}}
\usage{
bmi
}
\description{
A wide format data frame of simulated BMI values of 10,000 individuals.
}
\keyword{datasets}
