% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment.R
\name{class_assignment}
\alias{class_assignment}
\title{Maximum class assignment
\code{class_assignment}
Maximum class assignment}
\usage{
class_assignment(p)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}
}
\value{
This function computes the Average Posterior Probability Assignment (APPA) for a K latent class trajectory model.
}
\description{
Maximum class assignment
\code{class_assignment}
Maximum class assignment
}
\examples{
\dontrun{class_assignment(p)}
}
