% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen.kappa.R
\name{cohen.kappa}
\alias{cohen.kappa}
\title{Kappa matrix}
\usage{
cohen.kappa(x, w = NULL, n.obs = NULL, alpha = 0.05, levels = NULL)
}
\arguments{
\item{x}{Either a two by n data with categorical values from 1 to p or a p x p table. If a data array, a table will be found.}
}
\value{
Unweighted kappa
}
\description{
Kappa matrix of cohen's kappa values
}
\examples{
data(bmi_long, package="LCTMtools") # NO EXAMPLE SUPPLIED
}
