% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{A confusion matrix}
\usage{
confusion_matrix(model1, model2)
}
\arguments{
\item{model1}{A fitted model from the lcmm R package (or from SAS passed through the SASmodelbuilder() function)}

\item{model2}{is the posterior probabilities of assignment of dimensions, K columns and N rows}

\item{name1}{optional paramter to pre-specify name of model}

\item{name2}{optional paramter to pre-specify name of model}
}
\value{
A confusion matrix between two models with the same number of classes
}
\description{
A confusion matrix
}
\examples{
\dontrun{
data(bmi_long, package="LCTMtools")
require(lcmm)
model1 <- hlme(BMI ~Age, mixture= ~Age, random= ~Age, nwg=T, ng=2, subject="ID", data=bmi_long[1:500, ])
model2 <- hlme(BMI ~Age, mixture= ~Age, random= ~1, nwg=F, ng=2, subject="ID", data=bmi_long[1:500, ])
confusion_matrix(model1, model2)}
}
