% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(p)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}
}
\value{
Entropy value between (0, infinity)
}
\description{
A global measure of uncertainty with values close to zero implying a good model. Entropy is a global measure of classification uncertainty, which takes into account all N × K posterior probabilities. The entropy of a model is defined as which takes values from [0,infinity), with higher values indicating a larger amount of uncertainty. Entropy values closest to 0 correspond to models with least classification uncertainty.
}
\examples{
\dontrun{entropy(p)}
}
