% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of coloures}
}
\value{
a vector of n colour codes
}
\description{
Emulate ggplot2 default colour palette
}
\examples{
gg_color_hue(2)
gg_color_hue(4)
plot(1:10, pch=12, col=gg_color_hue(10), lwd=30, xaxt="n", yaxt="n", ylab="", xlab="")
}
\references{
https://www.rdocumentation.org/packages/iprior/versions/0.7.1/topics/gg_colour_hue
}
