% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_matrix.R
\name{kappa_matrix}
\alias{kappa_matrix}
\title{Kappa matrix}
\usage{
kappa_matrix(ConfMatrix, acc = 2)
}
\arguments{
\item{ConfMatrix}{A confusion matrix made using confusion_matrix(model1, model2)}

\item{acc}{The accuracy of the results, defaults to 2 decimal places}
}
\value{
Unweighted and weighted Kappa value computed using the cohen.kappa() function from the psych R package
}
\description{
Kappa matrix of cohen's kappa values
}
\examples{
data(bmi_long, package="LCTMtools")
require(lcmm)
require(psych)
model1 <- hlme(BMI ~Age, mixture= ~Age, random= ~Age, nwg=TRUE, ng=2, subject="ID", data=bmi_long[1:500, ])
model2 <- hlme(BMI ~Age, mixture= ~Age, random= ~1, nwg=FALSE, ng=2, subject="ID", data=bmi_long[1:500, ])
ConfMatrix <- confusion_matrix(model1, model2)
kappa_matrix(ConfMatrix)
}
