% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCTMcompare.R
\name{LCTMcompare}
\alias{LCTMcompare}
\title{A model comparison toolkit.
\code{LCTMcompare}}
\usage{
LCTMcompare(modelA, modelB)
}
\arguments{
\item{modelA}{is the output from hlme() R model or model is the output of SASmodelbuilder(oe, os, op, of) passed through it}

\item{modelB}{the model to be compared which is the output from hlme() R model or model is the output of SASmodelbuilder(oe, os, op, of) passed through it}
}
\value{
A selection of model adequacy tests, including the APPA (average posterior probability of assignment), the OCC (odds of correct classification), entropy $E$, Relative entropy ($E_k$),
}
\description{
The function LCTMcompare gives a summary of comparison between fitted LCTM models.
}
\examples{
data(bmi_long, package='LCTMtools')
require(lcmm)
set.seed(999)
data(bmi_long, package = 'LCTMtools' )
# Use the hlme function from the 'lcmm' R package to fit a 2 class latent class trajectory model
model2classes <- lcmm::hlme(fixed = bmi ~ age + I(age^2),
                      mixture= ~ age,
                      random = ~ age,
                      ng = 2,
                      nwg = TRUE,
                      subject = "id",
                      data = data.frame(bmi_long[1:500, ] ))
# Compute model adequacy measures
LCTMtoolkit(model2classes)
# Compare with a 3 class model
model3classes <- lcmm::hlme(fixed = bmi ~ age + I(age^2),
                      mixture= ~ age,
                      random = ~ age,
                      ng = 3,
                      nwg = TRUE,
                      subject = "id",
                      data = data.frame(bmi_long[1:500, ] ))
LCTMtoolkit(model3classes)
LCTMcompare(model2classes, model3classes)
}
\references{
\url{https://bmjopen.bmj.com/content/8/7/e020683}
}
