% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCTMtoolkit.R
\name{LCTMtoolkit}
\alias{LCTMtoolkit}
\title{A toolkit which computes a selection of model adequacy tests
\code{LCTMtoolkit}}
\usage{
LCTMtoolkit(model)
}
\arguments{
\item{modelA}{the models to be compared which is the output from hlme() R model or model is the output of SASmodelbuilder(oe, os, op, of) passed through it}

\item{modelB}{the models to be compared which is the output from hlme() R model or model is the output of SASmodelbuilder(oe, os, op, of) passed through it}
}
\value{
A selection of model adequacy tests, including the APPA (average posterior probability of assignment), the OCC (odds of correct classification), entropy $E$, Relative entropy ($E_k$),
}
\description{
The function LCTMtoolkit computes a selection of model adequacy tests, including the APPA (average posterior probability of assignment), the OCC (odds of correct classification), entropy E, Relative entropy (E_k), odds of correct classification is the ratio of the odds of classification based on the maximum posterior probablity classification rule and the estimated class membership proportions (pi_k)
}
\examples{
data(bmi_long, package="LCTMtools")
library(lcmm)
model2class <- hlme(BMI ~Age, mixture= ~Age, random= ~Age,
nwg=TRUE, ng=2, subject="ID", data=bmi_long[1:500, ])
postprob(model2class)
LCTMtoolkit(model2class)
}
