% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCTMtools.R
\docType{package}
\name{LCTMtools}
\alias{LCTMtools}
\alias{LCTMtools-package}
\title{LCTMtools: A package for computing a number of Latent Class Trajectory Model tools for a given hlme() object or SAS model.}
\description{
The LCTMtools package provides two categories of important functions:
LCTMtools (to test a models adequacy) and LCTMcompare (to aid model selection).
}
\section{LCTMtools functions}{

The LCTMtools functions arw a selection of model adequacy tests for Latent Class Trajectory Models (LCTMs) which include the APPA (average posterior probability of assignment), the OCC (odds of correct classification), entropy, Relative entropy.
}

