% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch.R
\name{mismatch}
\alias{mismatch}
\title{Computes the mismatch of the posterior probabilities (mismatch=actual-estimated)}
\usage{
mismatch(p, pi)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}

\item{pi}{is the estimated proportion of class membership of length K}
}
\value{
The mismatch of posterior probabilities
}
\description{
Computes the mismatch of the posterior probabilities (mismatch=actual-estimated)
}
\examples{
\dontrun{mismatch(p, pi)}
}
