% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmlcr_to_lctm.R
\name{mmlcr_to_lctm}
\alias{mmlcr_to_lctm}
\title{Converts an R mmclr model output to the format of R's hlme class}
\usage{
mmlcr_to_lctm(model)
}
\arguments{
\item{model}{contains model parameter estimates and maximised likelihood, AIC, BIC values}
}
\value{
A format to feed into the LCTMtoolkit() R function
}
\description{
Converts an R mmclr model output to the format of R's hlme class
}
\examples{
\dontrun{mmlcr_to_lctm(model)}
}
