% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.R
\name{occ}
\alias{occ}
\title{The odds of correct classification is the ratio of the odds of classification based on the maximum posterior probablity classification rule and the estimated class membership proportions (pi_k).}
\usage{
occ(p, pi)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}

\item{pi}{is the estimated proportion of class membership of length K}
}
\value{
The odds of correct classification
}
\description{
The odds of correct classification is the ratio of the odds of classification based on the maximum posterior probablity classification rule and the estimated class membership proportions (pi_k).
}
\examples{
\dontrun{occ(p, pi)}
}
