% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_entropy.R
\name{relative_entropy}
\alias{relative_entropy}
\title{The Relative Entropy}
\usage{
relative_entropy(p)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}
}
\value{
Relative Entropy - where values close to 1 indicate lowest classification uncertainty. In the special case when there is most uncertainty and each individual has equal probability of belonging to each class, E_K=0. Jedidi et al., describes relative entropy as a relative measure of ‘fuzziness’, and suggested cause concern when close to zero, as this implies that the latent class centroids are not sufficiently separated.
}
\description{
The Relative Entropy
}
\examples{
\dontrun{relative_entropy(p)}
}
