% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sastraj_to_lctm.R
\name{sastraj_to_lctm}
\alias{sastraj_to_lctm}
\title{Converts a SAS proc traj model to the format of R's hlme class}
\usage{
sastraj_to_lctm(oe, of, op, os)
}
\arguments{
\item{oe}{contains model parameter estimates and maximised likelihood, AIC, BIC values}

\item{of}{contains posterior probabilities}

\item{op}{contains predictors}

\item{os}{containts fixed effect  and class membershop parameter estimates}
}
\value{
A format to feed into the LCTMtoolkit() R function
}
\description{
Converts a SAS proc traj model to the format of R's hlme class
}
\examples{
\dontrun{sastraj_to_lctm(oe, of, op, os)}
}
