% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appa.R
\name{appa}
\alias{appa}
\title{Computes the Average Posterior Probability Assignment (APPA) for a K latent class trajectory model.}
\usage{
appa(p)
}
\arguments{
\item{p}{is the posterior probabilities of assignment of dimensions, K columns and N rows}
}
\value{
The Average Posterior Probability Assignment (APPA) for each class
}
\description{
Computes the Average Posterior Probability Assignment (APPA) for a K latent class trajectory model.
}
\examples{
\dontrun{appa(p)}
}
