% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_long.R
\docType{data}
\name{bmi_long}
\alias{bmi_long}
\title{Body Mass Index (BMI) repeated measures of 10,000 individuals in long format.}
\format{A list of two formats data frame with 53940 rows and 10 variables:
\describe{
  \item{id}{Individual ID}
  \item{age}{Age of BMI measure, in years}
  \item{bmi}{Body mass index of the individual at times T1,T2, T3 and T4, in kg/m^2}
  \item{true_class}{Tag to identify the class the individual BMI data was simulated from}
}}
\usage{
bmi_long
}
\description{
A list containing a long and wide format dataset of BMI values of 10,000 individuals.
}
\keyword{datasets}
