% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCTMcompare.R
\name{LCTMcompare}
\alias{LCTMcompare}
\title{A model comparison toolkit.
\code{LCTMcompare}}
\usage{
LCTMcompare(modelA, modelB)
}
\arguments{
\item{model}{is the output from hlme() R model or model is the output of SASmodelbuilder(oe, os, op, of) passed through it}
}
\value{
A selection of model adequacy tests, including the APPA (average posterior probability of assignment), the OCC (odds of correct classification), entropy $E$, Relative entropy ($E_k$),
}
\description{
The function LCTMcompare gives a summary of comparison between fitted LCTM models.
}
\examples{
data(bmi_long, package="LCTMtools")
require(lcmm)
modelA <- hlme(BMI ~Age, mixture= ~Age, random= ~Age,
nwg=TRUE, ng=2, subject="ID", data=bmi_long[1:500, ])
postprob(model2class)
modelB <- hlme(BMI ~Age, mixture= ~1, random= ~1,
nwg=FALSE, ng=2, subject="ID", data=bmi_long[1:500, ])
postprob(modelA, modelB)
LCTMcompare(modelA, modelB)
}
